requirements.txt
python-chess
gym